fix: log error for snapshots added
This is a bandaid for now so that the perf team can work on compaction,
but currently we don't seem to have anyone subscribed to the CompactedData
type's broadcast channel. When it sends a message it fails with an error.
To avoid killing the thread we just log the error for now. Follow up work
will try to fix this issue.