docs: cleanup catalog pro code docs and var names
This also addresses a logic error in the `DatabaseSchema::merge` method
that was not mapping tables that do not get updated, when there are others
that do get updated, for an existing db schema.

A test was added to check this change and fix any insta snapshots that
were affected.