feat: catalog merge and WAL mapping failure on incompatible last cache
Re-worked the catalog op mapping to produce errors when the
last cache create and delete ops are problematic.

Added a test to check for failure during catalog merge as well
as a test to check for failure during wal op mapping. In addition
a test was added that checks for a scenario where the replica
WAL is corrupted and contains a last cache create for a table
that does not exist.