fix: wal replication fails for field additions on non-existent table
During WAL replication, if a `AddFields` operation is encountered that
is for a table that does not exist, this will produce an error. This is
consistent with how operations are applied to a catalog on a primary.

A test was added to verify this behaviour.

The new helpers from the `influxdb3_wal` crate are used in the `replica.rs`
module as part of this PR.