feat: Switch FileIndex to use u64 not String
This changes up the FileIndex to use hashed u64 values of Strings
instead of using Strings. Lookup and inserts still use &str as this is
the boundry needed to determine in a query whether the column and value
is in the index. We now do not need to hold the name and the hashed value
in our FileIndex in memory.rs as well as using u64s for the actual
FileIndex in influxdb3_pro_index. We reuse the SerdeVecMap from
influxdb3_ids so that serialized and deserialized forms of the index are
now numbers rather than strings of numbers for efficiency, especially
since a u64 value would be a longer string in most cases.

Closes #129