feat: Update compactor with producer and consumer
This is a big refactor of the compactor that enables the multi-server configurations. It splits operation into a producer and a consumer and keeps a catalog for the compacted data, which is a union of the source host catalogs.

There are doc comments in compactor lib.

More tests are needed in the producer, compacted data, and consumer. Will log follow up issues for those.

We will also need to add startup flags that will enable a server to be a compactor without it having to be a read replica consuming wals. The compaction producer supports this but serve.rs currently doesn't have the configuration options to enable it.