fix: use placeholder queryexecutor in compactor mode
When running in compactor mode, creating default query executor relies
on calling methods `catalog()` on `Bufferer` trait that are
`unimplemented` which results in panic during startup. This commit
creates a placeholder query executor which implements `QueryExecutor`
and `QueryDatabase` trait and returns an error that is translated to
HTTP 405 Method not allowed error to client.

One thing to note is, because grpc setup involved passing in
`QueryDatabase`, although `QueryExecutor` extends `QueryDatabase` we
need to upcast it manually as now it's behind a `dyn` (trait object).
This requires a manual clone. On the surface it looks ok, as it is a one
time cost at the point of starting up.

closes: https://github.com/influxdata/influxdb_pro/issues/206