test: add e2e tests for meta cache create and delete REST API
Added an E2E test to verify success/failure modes of the metadata cache
creation REST API. Some of the internal error handling was modified so
that errors from the cache could be propagated accordingly.

Added an E2E test to verify success/failure modes of the metadata cache
deletion REST API. Error handling was updated for the only special
failure mode of the delete API, which is when the cache does not exist.

Added an E2E test to check that the metadata cache can be queried through
the REST API.