feat: remove old files during compaction
This commit changes the behavior of the compactor to remove uneeded
files while compacting. In particular it cleans up:

- Compaction Summaries
- Compaction Details
- Generation Details
- Uneeded compacted parquet files

It does so by keeping track of which files to delete as compaction
proceeds. When the compaction is completed a CompactionCleaner is
spawned with a delay of 10 minutes (this is changeable in tests). When
the 10 minutes is up, it will then delete each of the paths from object
storage as it is no longer needed.

Closes #163