feat: LastCacheExec to track predicate pushdown in last cache queries
The Predicate type was refactored to only support a IN or NOT IN type of
clause, to align with DataFusion's LiteralGuarantee analyzer.

This caused some churn in the last cache tests but they remain unchanged
in their assertions.

Added a test to check predicate handling from datafusion.

Updated code docs and comments to explain the changes.