feat: compaction consumer reload compacted catalog
This updates the compacted data consumer so that it will reload the compacted data catalog if it sees in the compaction summary that a new catalog was persisted.

A bit of the compacted catalog had to be refactored in order to enable this reloading in the running system.

In order to test the behavior, I ended up having to refactor the producer a bit. The benefit is that it's exercising a good amount of the end-to-end code path for compaction.

Fixes #262