fix: replicating catalog ops sends to the WAL in read_write
This refactors the WAL operation mapping in the influxdb3_catalog crate
for pro such that servers can replay and map WAL operations from another
host, but in addition, will send any CatalogOps received to their local
WAL. This will allow HA configurations.

Care was needed so that CatalogOps which would not result in a change
are ignored, and thereby not sent to the WAL, as doing so would create
an infinite loop between two or more hosts in an HA configuration.