fix(influxd): remove client context from wal write call (#614)
* fix(influxd): remove client context from wal write call

This cs changes the wal write call within writemulti to use a background
context. It no longer uses the client or customer's context as by the
time the wal write starts, the write to the cache has already completed
successfully. To avoid entropy sources, we don't want the
customer/client to be able to cancel the write to the wal. The wal write
can still fail for disk reasons e.g. Additionally, we disable the wal
write timeout; the default in oss is 10minutes which is mightly long,
but this cs makes it infinite. The wal write concurrency limiter is
still in place with a default of 2x go max procs.

* fix(influxd): change WALMaxWriteDelay default to 1 hour

We think we should have _some_ value here because the concurrency
limiter is also enabled - an hour though might as well be infinite for a
disk write.

* fix: actaully validate engine params

It turns out the validate function wasn't being called anywhere outside
tests. Now some engine params are validated when creating a new engine -
this means creating an engine can fail so that call now returns an error
and all the call sites have been updated.

(cherry picked from commit 4afc0be9d84eaf213555a132687be7f7db443a8b)