feat: Implement Deserialize for Catalog
This was fairly easy all things considered. We needed to turn on the
serde feature for parking_lot, remove the Serialize impl, derive both it
and Deserialize, and add the flatten attribute and we're all set. The
output for the tests changed, but only the order of fields, not the
actual content so it's fine that we updated them. This change will allow
us to deserialize the Catalog in the future if need be without running
into issues around the fact that it can't implement it.

Closes #25574