feat: create db on table create
This adds a reproducer for an issue where, in a multi-node setup, if a db
was created on one host, then a table for that db on another, the second
action would fail if the second host had not yet replicated the db create
from the first.

This also fixes said reproducer by making the table create also create the
database if it does not yet exist.