feat: Implement license loading and validation
This commit does two things:
1. It will load or create a license in the object store to be checked
   for validity.
2. It will validate that said license is not being used past the expiry
   date and spawn a task to continue validating that said license is
   valid
3. Disable querying endpoints so long as the license is expired

Note that:

1. This does work from local testing, but does not contain tests.
This is due to the fact that mocking it is hard to do without a valid
JWT token. For testing we will want to create a JWT for when a license
is expired or valid in the future.
2. This code will not work when run now unless you have a valid license,
create one locally with the license service, or you compile the code
with the `no-license` feature flag.

Closes #57
Closes #58
Closes #61