chore: Rename Pro to Enterprise
This commit renames anything with pro to enterprise instead. In some
cases like the README explicit references to Core are left around, but
in terms of user facing impacts the biggest is Enterprise instead of
Core in the CLI. Most of this is either renaming files or changing names
around so that things can compile.

Closes #343
Closes #330
Closes #344