refactor: rename --replicas to --read-from-writer-ids
Update the CLI argument --replicas to --read-from-writer-ids to be
consistent with the naming in core, i.e., --writer-id, which specifies
the writer identifier prefix of a write buffer being replicated.

This also changes the namign of metrics and logs that were using
"from_host" to use "from_writer" and updates code variable naming and
comments to use the term 'writer' instead of 'host' in the replica
code, where appropriate.

Co-authored-by: Paul Dix <paul@pauldix.net>